updated user specific routes to check for session data

+++ flask_app/flask06.py
@@ -29,9 +29,11 @@ with app.app_context():
 @app.route('/')
 @app.route('/index')
 def index():
-    a_user = db.session.query(User).filter_by(email='dshelt10@uncc.edu').one()
+    #check if a user is saved in session
+    if session.get('user'):
+        return render_template("index.html", user=session['user'])
+    return render_template("index.html")
 
-    return render_template('index.html', user = a_user)
 
 
 @app.route('/notes')
@@ -54,64 +56,72 @@ def get_note(note_id):
 
     return render_template('note.html', note = my_note, user=a_user)
 
+
+
 @app.route('/notes/new', methods=['GET', 'POST'])
 def new_note():
-    a_user = {'name': 'Duane', 'email':'dshelt10@uncc.edu'}
-    
-    if request.method == 'POST':
-        title = request.form['title']
 
-        text = request.form['noteText']
+    if session.get('user'):    
+        if request.method == 'POST':
+            title = request.form['title']
 
-        from datetime import date
-        today = date.today()
+            text = request.form['noteText']
 
-        today = today.strftime("%m-%d-%Y")
+            from datetime import date
+            today = date.today()
 
-        new_record = Note(title, text, today)
-        db.session.add(new_record)
-        db.session.commit()
+            today = today.strftime("%m-%d-%Y")
 
-        return redirect(url_for('get_notes', user = a_user))
+            new_record = Note(title, text, today, session['user_id'])
+            db.session.add(new_record)
+            db.session.commit()
 
-    else:  
-        a_user = db.session.query(User).filter_by(email='dshelt10@uncc.edu').one()
-        return render_template('new.html', user=a_user)
+            return redirect(url_for('get_notes'))
+
+        else:  
+            return render_template('new.html', user=session['user'])
+    else:
+        return redirect(url_for('login'))
 
 @app.route('/notes/edit/<note_id>', methods=['GET', 'POST'])
 def update_note(note_id):
-    #Check method for used request
-    if request.method == 'POST':
+    if session.get('user'):
+        #Check method for used request
+        if request.method == 'POST':
 
-        title = request.form['title']
+            title = request.form['title']
 
-        text = request.form['noteText']
-        note = db.session.query(Note).filter_by(id=note_id).one()
+            text = request.form['noteText']
+            note = db.session.query(Note).filter_by(id=note_id).one()
 
-        note.title = title
-        note.text = text
+            note.title = title
+            note.text = text
 
-        db.session.add(note)
+            db.session.add(note)
+            db.session.commit()
+
+            return redirect(url_for('get_notes'))
+        else:
+            # GET request - show new note form to edit note
+
+            #retrieve note from database
+            my_note = db.session.query(Note).filter_by(id=note_id).one()
+
+            return render_template('new.html', note=my_note, user=session['user'])
+    else:
+        return redirect(url_for('login'))
+
+        
+@app.route('/notes/delete/<note_id>', methods=['POST'])
+def delete_note(note_id):
+    if session.get('user'):
+        my_note = db.session.query(Note).filter_by(id=note_id).one()
+        db.session.delete(my_note)
         db.session.commit()
 
         return redirect(url_for('get_notes'))
     else:
-    # GET request - show new note form to edit note
-        #retrieve user from database
-        a_user = db.session.query(User).filter_by(email='dshelt10@uncc.edu').one()
-
-        #retrieve note from database
-        my_note = db.session.query(Note).filter_by(id=note_id).one()
-
-        return render_template('new.html', note=my_note, user=a_user)
-@app.route('/notes/delete/<note_id>', methods=['POST'])
-def delete_note(note_id):
-
-    my_note = db.session.query(Note).filter_by(id=note_id).one()
-    db.session.delete(my_note)
-    db.session.commit()
-
-    return redirect(url_for('get_notes'))
+        return redirect(url_for('login'))
 
 @app.route('/register', methods=['POST', 'GET'])
 def register():
